localrun/requirements.txt
# requirements.txt - Simplified Polling Architecture

# === Django Core ===
django
python-dotenv

# === API & Web ===
djangorestframework
django-cors-headers
djangorestframework-simplejwt

# === Celery & Results ===
celery
redis
django-celery-results

# === Pipeline Worker Dependencies ===
pandas
openpyxl
Pillow
pdf2image
together
requests


